baaten/requirements.txt
#google-search-results
#faiss-cpu
#sentence_transformers
#langchain_community
#streamlit
#youtube_search
#langchain
#PyPDF2
#chromadb
#langchain_google_genai
#langdetect
#langchain==0.1.16
#langchain-community==0.0.33
#langchain-google-genai==0.0.8
langchain==0.1.16
langchain-community==0.0.33
langchain-google-genai==0.0.8
google-generativeai==0.3.2
googletrans==4.0.0-rc1
faiss-cpu
streamlit
python-dotenv
PyPDF2
langdetect
pymupdf
rank-bm25
beautifulsoup4>=4.12.2
openai

#C:\Users\anshu\Desktop\Ajit_group\ajit